requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
pandas==2.2.2
pydantic==2.9.2
python-multipart==0.0.9
requests==2.32.3
python-dotenv
httpx